Replace deprecated climate.CLIMATE_SCHEMA
When compiling we get a warning about the deprecation of
climate.CLIMATE_SCHEMA:

WARNING Using `climate.CLIMATE_SCHEMA` is deprecated and will be removed in ESPHome 2025.11.0. Please use `climate.climate_schema(...)` instead. If you are seeing this, report an issue to the external_component author and ask them to update it. https://developers.esphome.io/blog/2025/05/14/_schema-deprecations/. Component using this schema: hisense_ac

This commit fixes #3